requirements.txt
telebot
pyTelegramBotAPI
mysql.connector
